moving auth action provider to its own file
